interlibrary_loan_request
This is the javascript to create a ILL (interlibrary loan request) button in Primo brief display and full display results. When the patron clicks the request button, it will populate and create a ILL button so that the patron has the choice to go to iLLiad to launch ILL request.